Remove a bunch of NPA files that have been superceded
And copy most current NPA_metrics_Goal_3_v2.py to NPA_metrics_Goal_3.py
